+++ CHANGELOG.md
@@ -3848,7 +3848,7 @@ from functools import partial to all files and makes the docs cleaner by renderi
 - rename Component `to_dict` to `to_dict()`: is now a method instead of a property
 - rename Component `pprint` to `pprint()`: is now a method instead of a property
 - rename Component `pprint_ports` to `pprint_ports()`: is now a method instead of a property
-- Component.mirror() returns a container
+- Component.dmirror() returns a container
 
 ## 3.4.4